// frontend/src/api/generated/models/stocktake-line-item.ts
/* tslint:disable */
/* eslint-disable */
/**
 * WMS ShowCase API
 * 倉庫管理システム（WMS）ShowCase のREST API定義
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


// May contain unused imports in some cases
// @ts-ignore
import type { UnitType } from './unit-type';

/**
 * 
 * @export
 * @interface StocktakeLineItem
 */
export interface StocktakeLineItem {
    /**
     * 棚卸明細ID
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'lineId': number;
    /**
     * ロケーションID
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'locationId': number;
    /**
     * ロケーションコード
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'locationCode': string;
    /**
     * 商品ID
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'productId': number;
    /**
     * 商品コード
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'productCode': string;
    /**
     * 商品名
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'productName': string;
    /**
     * 
     * @type {UnitType}
     * @memberof StocktakeLineItem
     */
    'unitType': UnitType;
    /**
     * ロット番号
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'lotNumber'?: string | null;
    /**
     * 賞味/使用期限
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'expiryDate'?: string | null;
    /**
     * 棚卸開始時在庫数
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'quantityBefore': number;
    /**
     * 実数（null=未入力）
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'quantityCounted'?: number | null;
    /**
     * 差異数（null=棚卸確定前）
     * @type {number}
     * @memberof StocktakeLineItem
     */
    'quantityDiff'?: number | null;
    /**
     * 実数入力済みフラグ
     * @type {boolean}
     * @memberof StocktakeLineItem
     */
    'isCounted': boolean;
    /**
     * 実数入力日時
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'countedAt'?: string | null;
    /**
     * 実数入力者氏名
     * @type {string}
     * @memberof StocktakeLineItem
     */
    'countedByName'?: string | null;
}